// WADNR.Web/src/app/shared/generated/model/fund-source-allocation-program-index-project-code-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FundSourceAllocationProgramIndexProjectCodeItem { 
    FundSourceAllocationProgramIndexProjectCodeID?: number;
    ProgramIndexID?: number;
    ProgramIndexCode?: string | null;
    ProjectCodeID?: number | null;
    ProjectCodeName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FundSourceAllocationProgramIndexProjectCodeItemForm { 
    FundSourceAllocationProgramIndexProjectCodeID?: FormControl<number>;
    ProgramIndexID?: FormControl<number>;
    ProgramIndexCode?: FormControl<string>;
    ProjectCodeID?: FormControl<number>;
    ProjectCodeName?: FormControl<string>;
}

export class FundSourceAllocationProgramIndexProjectCodeItemFormControls { 
    public static FundSourceAllocationProgramIndexProjectCodeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramIndexID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramIndexCode = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectCodeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectCodeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}